refactor(web/applications): move session banner code to common location
