+++ apps/web/src/server/applications/common/services/session.service.js
@@ -154,3 +154,41 @@ export const getSuccessBanner = (session) => {
 export const destroySuccessBanner = (session) => {
 	delete session.showSuccessBanner;
 };
+
+/**
+ * Session data is used for the success banner on the project updates page and overview page
+ * It is deleted once read, as it only applies once
+ *
+ * @typedef {import('express-session').Session & { banner?: string }} BannerSession
+ */
+
+/**
+ * Get the banner for the user from the session.
+ *
+ * @param {BannerSession} session
+ * @returns {string|undefined}
+ */
+export function getSessionBanner(session) {
+	return session.banner;
+}
+
+/**
+ * Set the banner for the user in the session.
+ *
+ * @param {BannerSession} session
+ * @param {string} banner
+ * @returns {void}
+ */
+export function setSessionBanner(session, banner) {
+	session.banner = banner;
+}
+
+/**
+ * Delete the banner for the user from the session.
+ *
+ * @param {BannerSession} session
+ * @returns
+ */
+export function deleteSessionBanner(session) {
+	delete session.banner;
+}
